% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spec.R
\name{build_spec}
\alias{build_spec}
\title{Convert function args to a dataframe}
\usage{
build_spec(...)
}
\arguments{
\item{...}{A combination of named arguments, which will each become a column,
and unnamed arguments, which will be combined into a column called \code{datapoint}}
}
\description{
Convert function args to a dataframe
}
