% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ec_as_js.ec_object}
\alias{ec_as_js.ec_object}
\title{TODO: Handle Javascript properly so it's not a string in the final result}
\usage{
\method{ec_as_js}{ec_object}(x)
}
\description{
TODO: Handle Javascript properly so it's not a string in the final result
}
