% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{ec_as_js}
\alias{ec_as_js}
\title{Generate JS version
TODO: Handle Javascript properly so it's not a string in the final result}
\usage{
ec_as_js(option)
}
\description{
Generate JS version
TODO: Handle Javascript properly so it's not a string in the final result
}
