% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_series.R
\name{ec_bar}
\alias{ec_bar}
\title{Bar series}
\usage{
ec_bar(df, x, y, ...)
}
\arguments{
\item{df}{A dataframe that can be referenced via ... arguments}

\item{x}{The x-axis variable}

\item{y}{The y-axis variable}

\item{...}{Additional arguments to be passed to the series}
}
\description{
https://echarts.apache.org/en/option.html#series-bar
}
\details{
Generates a single bar series or list of series from a dataframe.
}
\examples{

# Including x-value in series data
data <- tibble::tibble(x = 1:5, y = rnorm(5))
series <- ec_bar(data, x, y)

echartr(option = list(
  xAxis = list(type = "value"),
  yAxis = list(type = "value"),
  legend = list(show = TRUE),
  series = series
))

# Including x-value in axis data - requires category type axis
data <- tibble::tibble(x = 1:5, y = rnorm(5))
series <- ec_bar(data, y = y)

echartr(option = list(
  xAxis = list(type = "category", data = data$x),
  yAxis = list(type = "value"),
  legend = list(show = TRUE),
  series = series
))

}
