% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ec_clip.ec_object}
\alias{ec_clip.ec_object}
\title{Copy the javascript representation of chart options}
\usage{
\method{ec_clip}{ec_object}(x)
}
\arguments{
\item{x}{echartr list-tree}
}
\description{
Copy the javascript representation of chart options
}
