% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ec_clip}
\alias{ec_clip}
\title{Copy the javascript equivalent of an echartr list-tree object to the clipboard}
\usage{
ec_clip(x)
}
\arguments{
\item{x}{echartr list-tree}
}
\description{
Copy the javascript equivalent of an echartr list-tree object to the clipboard
}
