% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_data.R
\name{ec_data}
\alias{ec_data}
\title{Generate echarts \code{data} list-tree object from a dataframe}
\usage{
ec_data(df, ...)
}
\arguments{
\item{df}{a dataframe that can be referenced by all other arguments}

\item{...}{data dimensions and attributes.

Unnamed arguments are used as data dimensions.
Named arguments should be valid data attributes.}
}
\description{
Generate echarts \code{data} list-tree object from a dataframe
}
