% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_data.R
\name{ec_data_}
\alias{ec_data_}
\title{Generate echarts \code{data} list-tree object from specifying dataframe}
\usage{
ec_data_(spec)
}
\arguments{
\item{spec}{specification data frame}
}
\description{
Generate echarts \code{data} list-tree object from specifying dataframe
}
\keyword{internal}
