% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_series.R
\name{ec_line}
\alias{ec_line}
\title{Line series}
\usage{
ec_line(df, ...)
}
\arguments{
\item{df}{A dataframe that can be referenced via ... arguments}

\item{...}{Series attributes}
}
\description{
https://echarts.apache.org/en/option.html
}
\details{
Adds option attribute \code{type = "line"} and will give an error if \code{y} is not provided.
}
\examples{

# Including x-value in series data
data <- tibble::tibble(x = 1:5, y = rnorm(5))
series <- ec_line(data, x, y)

echartr(option = list(
  xAxis = list(type = "value"),
  yAxis = list(type = "value"),
  legend = list(show = TRUE),
  series = series
))

# Including x-value in axis data - requires category type axis
data <- tibble::tibble(x = 1:5, y = rnorm(5))
series <- ec_line(data, y)

echartr(option = list(
  xAxis = list(type = "category", data = data$x),
  yAxis = list(type = "value"),
  legend = list(show = TRUE),
  series = series
))

}
\seealso{
\code{\link[=ec_series]{ec_series()}}
}
