% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_options.R
\name{ec_options}
\alias{ec_options}
\title{Generate a list of options from a dataframe}
\usage{
ec_options(df, ...)
}
\arguments{
\item{df}{A dataframe that can be referenced via ... arguments}

\item{...}{Expressions that can be evaluated to generate the list-tree}
}
\description{
Generate a list of options from a dataframe
}
