% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_options.R
\name{ec_options_}
\alias{ec_options_}
\title{Create a list of options based on specification data frame}
\usage{
ec_options_(spec)
}
\description{
Create a list of options based on specification data frame
}
