% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_series.R
\name{ec_series}
\alias{ec_series}
\title{Generate a list of series from a dataframe}
\usage{
ec_series(df, ...)
}
\arguments{
\item{df}{A dataframe that can be referenced by all other arguments}

\item{...}{data dimensions and series attributes

Unnamed arguments are used as data dimensions in the order provided.
Named arguments should be valid series attributes.}
}
\description{
Generate a list of series from a dataframe
}
\seealso{
\code{\link[=ec_scatter]{ec_scatter()}}

\code{\link[=ec_line]{ec_line()}}

\code{\link[=ec_bar]{ec_bar()}}

\code{\link[=ec_pie]{ec_pie()}}
}
