% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ec_object}
\alias{print.ec_object}
\title{Print the javascript representation of an echartr list-tree object}
\usage{
\method{print}{ec_object}(x)
}
\arguments{
\item{x}{S3 object with class "ec_object"}
}
\description{
Print the javascript representation of an echartr list-tree object
}
