% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_eval}
\alias{row_eval}
\title{Evaluate quosures provided to ... for each for in df}
\usage{
row_eval(df, ...)
}
\arguments{
\item{df}{A dataframe to evaluate the quosures against}

\item{...}{A combination of named arguments, which will each become a column}
}
\description{
Evaluate quosures provided to ... for each for in df
}
\keyword{internal}
