% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spec.R
\name{spec_zoom}
\alias{spec_zoom}
\title{Move one level 'deeper' into the specification
TODO: support unnamed stuff inside row (maybe)}
\usage{
spec_zoom(spec, col)
}
\description{
Move one level 'deeper' into the specification
TODO: support unnamed stuff inside row (maybe)
}
