% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartr.R
\name{updateEchartr}
\alias{updateEchartr}
\title{Intention is to support all functions under https://echarts.apache.org/en/api.html#echartsInstance}
\usage{
updateEchartr(
  session = shiny::getDefaultReactiveDomain(),
  outputId,
  option = NULL,
  on = NULL,
  off = NULL
)
}
\description{
Intention is to support all functions under https://echarts.apache.org/en/api.html#echartsInstance
}
