% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ec_as_js}
\alias{ec_as_js}
\title{Convert an echartr list-tree object into javascript}
\usage{
ec_as_js(x)
}
\arguments{
\item{x}{echartr list-tree}
}
\description{
Convert an echartr list-tree object into javascript
}
