% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_data.R
\name{ec_data}
\alias{ec_data}
\title{Generate echarts \code{data} list-tree object from a dataframe}
\usage{
ec_data(df, ...)
}
\arguments{
\item{df}{data frame that can be referenced via ... arguments}

\item{...}{Expressions that can be evaluated to generate the list-tree}
}
\description{
Generate echarts \code{data} list-tree object from a dataframe
}
