% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_series.R
\name{ec_line}
\alias{ec_line}
\title{Line series}
\usage{
ec_line(df, x, y, ...)
}
\description{
Line series
}
\examples{

# Including x-value in series data
data <- tibble::tibble(x = 1:5, y = rnorm(5))
series <- ec_line(data, x, y)

echartr(option = list(
  xAxis = list(type = "value"),
  yAxis = list(type = "value"),
  legend = list(show = TRUE),
  series = series
))

# Including x-value in axis data - requires category type axis
data <- tibble::tibble(x = 1:5, y = rnorm(5))
series <- ec_line(data, y = y)

echartr(option = list(
  xAxis = list(type = "category", data = data$x),
  yAxis = list(type = "value"),
  legend = list(show = TRUE),
  series = series
))

}
