% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_options.R
\name{ec_options}
\alias{ec_options}
\title{Generate a list of options from a dataframe}
\usage{
ec_options(df, ...)
}
\arguments{
\item{df}{a dataframe that can be referenced by all other arguments}

\item{...}{data dimensions and option attributes.

Unnamed arguments are used as data dimensions.
Named arguments should be valid option attributes.}
}
\description{
Generate a list of options from a dataframe
}
