% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_series.R
\name{ec_scatter}
\alias{ec_scatter}
\title{Scatter series}
\usage{
ec_scatter(df, x, y, ...)
}
\description{
Generates a single scatter series or list of series from a dataframe.
}
\details{
https://echarts.apache.org/en/option.html#series-scatter
}
