% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ec_option}
\alias{print.ec_option}
\title{Print the javascript representation of an echartr list-tree object}
\usage{
\method{print}{ec_option}(x)
}
\description{
Print the javascript representation of an echartr list-tree object
}
