% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_eval}
\alias{row_eval}
\title{Evaluate quosures provided to ... for each row in a dataframe}
\usage{
row_eval(df, ...)
}
\arguments{
\item{df}{A dataframe to evaluate the quosures against}

\item{...}{A combination of named arguments, which will each become a column

Unnamed arguments are combined as lists into a column called "unnamed"}
}
\description{
Evaluate quosures provided to ... for each row in a dataframe
}
\keyword{internal}
