% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spec.R
\name{spec_zoom}
\alias{spec_zoom}
\title{Move one level 'deeper' into the specification}
\usage{
spec_zoom(spec, col)
}
\arguments{
\item{spec}{specification data frame}

\item{col}{character, column name to zoom in on}
}
\description{
Move one level 'deeper' into the specification
}
\keyword{(maybe)}
\keyword{TODO:}
\keyword{inside}
\keyword{internal}
\keyword{row}
\keyword{stuff}
\keyword{support}
\keyword{unnamed}
